Server error experiment

+++ app/src/main/java/com/example/individual/data/repository/PlaneRepository.kt
@@ -32,8 +32,12 @@ class PlaneRepository {
     }
 
     suspend fun deleteById(planeId: Long) {
-        individualApi.deletePlane(planeId)
-        planeDao.deleteById(planeId)
+        // Если запрос возвращается с ошибкой, то корутина не умирает
+        // Как выводить сообщение об ошибке
+        val response = individualApi.deletePlane(123456789)
+        if (response.isSuccessful) {
+            planeDao.deleteById(planeId)
+        }
     }
 
     suspend fun getPlaneById(id: Long): PlaneFull {
